chore(rust): improve documents of `OxcAst`
